[FIX] invoice_xunnel: Search invoice related to credit note
Case:
Invoice with uuid '6A8F9B60-2B51-4F05-93B4-8BC418C3166C' (uppercased)

Credit note that contains in the cfdi node: CfdiRelated lowercase uuid
'6a8f9b60-2b51-4f05-93b4-8bc418c3166c'

When importing the xml of the credit note, does not get the
related invoice, because it does not compare upper and lower case for
uuid.

This fix changes the uuid of the invoices found to lowercase and the
uuid of the XML of the credit note to lowercase, to compare them.