[FIX] invoice_xunnel,account_xunnel: minor fixes
- The python of the wizard used to save the connection was refactored, because the default
onchange, of the related field does not changed the original value, also we had security
problems, because we need to modify the connection token in the main company.
- The view of the wizard used to save the connection was improved:
  * connection token use password widget
  * the default footer and buttons were replaced to a new one to ensure the token saved
- The python of the wizard used to download the XML from SAT refactored to avoid
using related fields, due to security problems when trying to modify the last_sync_date
of the current company.
- The message error when the Credential creator is opened was improved, using the actual
error.

This commit is related to the issue https://git.vauxoo.com/vauxoo/instance/-/issues/772